apps/api/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
openpyxl
python-multipart
numpy
scipy
python-docx
